requirements.txt
flask==1.0.2
python-dotenv==0.13.0
Pillow==7.1.2
flask_restx==0.2.0
gunicorn==20.0.4
flask-cors==3.0.8
requests==2.23.0
pymongo==3.10.1
dnspython==1.16.0
